Various build fixes with Qt6
- Fix some include directives
- Fix QtConcurrent::run usage, the signature has changed in Qt6
- Q6ByteArray::size() returens qsizetype, aka long long, adjust sprintf()
  syntax
- int can't be implicitly converted to QChar any more, create a QChar
  explicitly
- Q6StringBuilder is confused when concatenating QByteArray's, help it
- Q6FileInfo constructor that takes a QString is now explicit, use
  QDirIterator::fileInfo() directly
- QString::size() returns a qsizetype, cast it to int when creating a
  QJSValue
- Use QSortFilterProxyModel::filterRegularExpression(), QRegularExpression
  vs. QRegExp (deprecated in Qt6)
- Create a QKeySequence explicitly, implicit conversion from int is gone

GIT_SILENT

+++ src/kpac/script.cpp
@@ -295,7 +295,7 @@ public:
             return QJSValue(0);
         }
 
-        return QJSValue(host.count(QLatin1Char('.')));
+        return QJSValue(static_cast<int>(host.count(QLatin1Char('.'))));
     }
 
     // shExpMatch(str, pattern)
